Linux compilation fixes

+++ lmath/Geometry.h
@@ -82,12 +82,12 @@ inline vec3 getBarycentricCoordinates(const vec3* pts, vec3 p)
 
 LFORCEINLINE vec3 quantizeFloor(const vec3& v, float scale)
 {
-	return vec3(floor(v.x / scale) * scale, floor(v.y / scale) * scale, floor(v.z / scale) * scale);
+	return vec3(floorf(v.x / scale) * scale, floorf(v.y / scale) * scale, floorf(v.z / scale) * scale);
 }
 
 LFORCEINLINE vec3 quantizeCeil(const vec3& v, float scale)
 {
-	return vec3(ceil(v.x / scale) * scale, ceil(v.y / scale) * scale, ceil(v.z / scale) * scale);
+	return vec3(ceilf(v.x / scale) * scale, ceilf(v.y / scale) * scale, ceilf(v.z / scale) * scale);
 }
 
 } // namespace ldr
